MB-34654: Impose strict flusher batch limit on Disk Checkpoints
To prevent a large spike in memory usage that can push us over the
high watermark and cause us to return temp OOMs to clients, force
the splitting of flusher batches for Disk Checkpoints (i.e. backfills).
This lessens the front-end impact of large rebalances. Replicas can
already deal with warming up a partial snapshot and the behaviour has
not changed in this patch for active.

Change-Id: Id23fd4c432fdb6ff316a7ec3923cef560de7e9db
Reviewed-on: http://review.couchbase.org/121500
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>