MB-37748: Shared locking for bucket Collection::Manifest
Primary change was to change the Collection::Manger mutex to be
a shared lock so that concurrent readers don't have to serialise
around a read.

As part of that change is the removal of the "no-manifest" state,
the Manager will always construct a Manifest, and this Manifest
is in the initial collection state, that is:

    uid:0
    scopes:[_default] (1 scope)
    collections[_default] (1 collection)

The update code is also adjusted so that it doesn't hold exclusive
access on the current state whilst it is validating/applying the new
state, however the update function itself is exclusive.

The update also remains "paranoid", rather than accepting the manifest
blindly, we still apply some checks, e.g. uid is increasing and now
we expect that a uid:0 manifest as input, matches our current state.

Change-Id: I00013df2195d01a980625b85b1fb0b397f549735
Reviewed-on: http://review.couchbase.org/121575
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>