MB-36379: Add static limit for engine file descriptors to Environment
Take the remaining FDs that memcached did not allocate for desired
max number of connections (set via memcached.json) and allocate these to
the engines. For now, this means that the number of max_connections
cannot increase beyond the initial limit at runtime. In the future we
will allow the engines to have a dynamically changeable number of file
descriptors and this can change.

Add const values for the minimum number of file descriptors given to
the engines and the file descriptors reserved for things other than
connections to the Environment. Both of these numbers are relatively
arbitrary, but are small percentages of the current nofile ulimit we
attempt to set on the couchbase user of 70,000 per process.

Change-Id: I4ce193820cf143bf87b26c03926366e2d1b67e2e
Reviewed-on: http://review.couchbase.org/121666
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>