Make couchstore specific KVStoreConfig and move config to derived class
We have Magma and RocksDB specific KVStoreConfig classes, and the
KVStoreConfig class has couchstore specific members in it. Move these
to a new couchstore specific class.

Doing this requires additional dynamic casting in CouchKVStore to
convert between KVStoreConfig and CouchKVStoreConfig. Move the
ownership of the config to the derived KVStore class.

Change-Id: I2f6164697a0886f2d72ae40576d9043c433b8e12
Reviewed-on: http://review.couchbase.org/121734
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>