Build nobucket as a static library
We don't need nobucket to be a shared library. In theory all of our
engines could be linked statically as we no longer load them
dynamically at link time. By keeping some of them as shared objects
we get better integration with perf (as it can split the CPU time
into shared objects), but perf analysis have a limited value
for nobucket ;-)

Change-Id: I6d446527b98cbe12e80c2af03e4fb1d34519a12d
Reviewed-on: http://review.couchbase.org/121737
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>