Don't specify a prefix for ep and default_engine
Back in the days we had ns_server tell us the names for the
libraries to open and we tried to open them via calling dlopen
and LoadLibrary. Given that unix like operatins sytems used
a lib prefix and windows didn't use any at all we took the
easy way out and built them without the lib prefix so that
all we had left was to try to "guess" the file extension
(.dll, .dylib, .so etc).

Now that memcached is linked with the library and the
runtime linker do all the heavy lifting resolving this
we no longer need to force a special prefix on the library.

Change-Id: I013bf844e962040992720ebcb851b03f50703d58
Reviewed-on: http://review.couchbase.org/121743
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>