Reduce the dependency list for clang-tidy builds
When running in the context for clang-tidy CV we want to
build as little as possible. The programs we've got to generate
source files links with dirutils (which in turn links with platform,
which links with phosphor) just to read a file. Create a separate
method to do so to aviod linking with dirutils.

In addition inline cb::throwJsonTypeError so that we don't need
to build mcd_util just to throw the exception

Change-Id: I0963617f5182d7cc6d15a88c73fcb4e5f5833543
Reviewed-on: http://review.couchbase.org/121790
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>