MB-37772: Restructure DCPBackfill so we can have BySeqno/ById
This patch restructures the backfill classes so that we have
the following class hierarchy.

This structure will enable the creation of a new disk based
class and share relevant code from the current by-seqno
disk backfill code.

1) DCPBackfill (backfill.h/.cc) is the base-class

2) DCPBackfillBySeqno (backfill_by_seqno.h) common code/data for any
backfill that iterates a seqno index (not a concrete class).

    DCPBackfill
        | (virtual public)
    DCPBackfillBySeqno

3) DCPBackfillDisk (backfill_disk.h/.cc) common code/data for any
backfill that uses disk (not a concrete class).

    DCPBackfill
        | (virtual public)
    DCPBackfillDisk

With these 3 building blocks we can create the two backfill classes we
will use and in a future patch we can create a new disk based backfill.

1) DCPBackfillMemoryBuffered (backfill_memory.h/.cc) is the class used
for ephemeral backfills.

    DCPBackfill
        | (virtual public)
    DCPBackfillBySeqno
        | (public)
    DCPBackfillMemoryBuffered

2) DCPBackfillBySeqnoDisk (backfill_by_seqno_disk.h/.cc) is the class
used for DCP seqno based backfills for persistent buckets.

                 ___ DCPBackfill ___
                | (virtual public)  |
    DCPBackfillBySeqno       DCPBackfillDisk
                |     (public)      |
                DCPBackfillBySeqnoDisk

Change-Id: Id15432420986be5681f618e96292886f475c6559
Reviewed-on: http://review.couchbase.org/122404
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>