MB-37096: WinDebug: Fix IncompleteDiskSnapshotWarmsUpToHighSeqno failure
The Windows Debug build doesn't perform automatic copy elision in the
return from DurabilityWarmupTest::resetEngineAndWarmup(). As a result,
a temporary copy of PrePostStateChecker is created and destroyed during
the return.

This results in the test failing with:

    [ RUN      ] FullOrValue/DurabilityWarmupTest.IncompleteDiskSnapshotWarmsUpToHighSeqno/persistent_value_only
    ..\kv_engine\engines\ep\tests\module_tests\evp_store_warmup_test.cc(747): error: Value of: vb->getHighCompletedSeqno()
      Actual: 0
    Expected: preHCS
    Which is: 1
    PrePostStateChecker: Found that post warmup the HCS does not match the pre-warmup value

This is bcause the temporary copy of PrePostStateChecker has a valid
VBucketPtr, and hence the dtor incorrectly checks the pre HPS and HCS
(which this test has explicitly disabled):

    // HCS will go backwards (to 0) so we can't make the normal checks here;
    // this is because we flush it on snapshot end in the flusher and we have
    // not received the snapshot end. This is expected.
    resetEngineAndWarmup().disable();

Fix by disabling copy of PrePostStateChecker (it's not intended to be
copied).

Change-Id: Ic99b20b7429eb835e1d666f98acddffb82ce10d5
Reviewed-on: http://review.couchbase.org/121851
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>