MB-37827: Take shared_ptr to stream in producer seqno ack
There is a race condition as follows.

1) Bucket shutdown starts but has not yet destroyed streams of our
given producer.
2) A seqno ack comes in and gets partially processed. We find the
stream in the producer but not yet process the ack.
3) Bucket shutdown continues and destroys the stream object by
removing the owning shared_ptr in DcpProducer::closeAllStreams
4) Seqno ack processing continues and segfaults when attempting to
access the stream.

Fix this by taking a shared_ptr to the stream in
DcpProducer::seqno_acknowledged instead of a raw pointer to ensure
that it can't be deleted under us.

Change-Id: Ibfc6a82515b55f96acad5d5df5e39bf860ef08bb
Reviewed-on: http://review.couchbase.org/122050
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Well-Formed: Wayne Siu <wayne@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>