MB-37772: Refactor ScanContext so it owns the KVStore handle
This patch aligns Rocks/Couch KVStore with Magma, in that
instead of the ScanContext holding a 'scan-id' which is used
to lookup the 'handle' required for the scan(), we embed the
abstract KVFileHandle into ScanContext. Now ScanContext owns
the handle which will be useful for allowing backfill code to
further interrogate the KVStore (e.g. call getCollectionStats).

This also updates the KVFileHandle/makeFileHandle code to
remove the use of the custom deleter, it's no longer required.
Destruction of a KVFileHandle will trigger release of resources
as all subclasses of KVFileHandle employ RAII.

Change-Id: I708d663bfa44b8b93cdf9a88fd05e73f63517d00
Reviewed-on: http://review.couchbase.org/122049
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>