Fix Clang Static Analyzer warnings related to SingleThreadedRCPtr<>
Fix clang warning now being generated due to an upgrade form
scan-build-6.0 to scan-build-9. The warnings fixed in this patch are all
related to SingleThreadedRCPtr<>. To do this inline the ref counting so
that the analyzer can understand the what the state of RCValue is. For
robustness we will also exclude the clang analyzer from seeing the
Deleter() function in SingleThreadedRCPtr<> as in some cases it might
not track RCValue ref count correctly. Which should stop any false
positives from about SingleThreadedRCPtr<> in the future.
Also clean up a few clang-tidy warnings and suppress a clang analyzer
false positive in evp_store_single_threaded_test.cc.

Change-Id: Iffb11fe57078b9637a28bb170f3678ae96029941
Reviewed-on: http://review.couchbase.org/122059
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>