MB-37772: Include default collection in filter set
Prior to this change a non-passthrough filter would not bother
storing the default collection in the filter set, instead the
bool defaultAllowed is a 'cache' of that state.

However we now have some need to write code which needs to consider
a DCP streams filter set.

To simplify the external API we now add the default collection into
the set, changing a client's code from something like:

    if (filter.isDefaultAllowed()) {
        x.push_back(CollectionID::Default);
    }
    for (auto c : filter) {
        x.push_back(c);
    }

to a single loop:

    for (auto c : filter) {
        x.push_back(c);
    }

Note: that all replication streams are now passthrough streams,
only specialised collection streams would reach the deeper
checks inside the filter class.

Note: Future work may look to now remove the defaultAllowed cache
as now that KV is fully collection aware, it's less important to
have this special case.

Change-Id: I16ddb621177f239785a7e1e5bb81d91cf7e841b4
Reviewed-on: http://review.couchbase.org/122122
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>