MB-35388: Add HELLO::VAttr flag for graceful VAttr detection
Add a new HELLO flag which allows clients to (safely) check for the
availablity of Virtual Attributes.

Strictly speaking, Virtual Attributes (VATTRs) have been available
since v5.0 when $document was added. Later a second VATTR, $XTOC was
added in v5.5. However, no HELLO feature flag was added at these times.

However, it is complex for a client to programmatically check for VATTRs:

- There is no HELLO flag to check.

- Attempting to probe for a VATTR which may or may not exist ($XTOC,
  or the soon to be added $vbucket) was cumbersome as the server would
  disconnect the client if it attempted to access an unknown VATTR -
  e.g.

    WARNING 63: Invalid format specified for "SUBDOC_MULTI_LOOKUP" -
    Status: "Subdoc: Unknown xattr virtual attribute" - Closing
    connection.

  A client would therefore have to maintain state about the
  capabilites of a KV-Engine node across the lifetime of a conneciton
  object, which also introduces potential races if the actual node
  changed (e.g. rebalance) between disconnect and reconnect.

  Note this was arguably a bug in the original VATTR implementation,
  the intent was that the client would _not_ be disconnected but would
  receive the status code SubdocXattrUnknownVattr.

To address this limitation, essentially fix the original oversight -
defer the validation of VATTR paths to the actual execution phase
(instead of in packet validation phase). This means a client will no
longer be disconnected if they attempt to access an unknown VAttr,
instead they will receive the Status code SubdocXattrUnknownVattr
(0xd1) and the connection will not be affected.

To flag this new behaviour to the client, add a "new" HELLO::VAttr
feature flag. While this isn't technically a new feature, this seems a
sensible Hello flag name going forward (and I couldn't think of a
good, short alternative which otherwise conveyed this semantic change
- Hello::VATTRThisTimeWithCorrectMissingVAttrHandling ;)

Change-Id: I461d2e1bfa59ff723d9754609df76ae51601de4e
Reviewed-on: http://review.couchbase.org/122602
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>