MB-37881: Create correct number of shards in ep-engine_ep_unit_tests
Currently in ep-engine_ep_unit_tests we set the number of shards
by default to 2. However, we actually create the correct number of
shards based on the WorkLoadPolicy object passed to the EPStore.
This is currently hard coded to 1. Correct this to create the correct
number of shards.

This patch caused a few test failures. The flusher_test attempted to
find the correct flusher object but would find one for a different
shard as the vBuckets no longer map to the same shard. Multiple issues
were found with the collections dcp tests. Primarily the
CollectionsDcpTest::getPersistedManifest() function was incorrectly
always retrieving the collections manifest of vBucket 0. This meant
that comparisons were meaningless. The test now fails because vBucket 0
may be on a different shard that whatever Vbid is passed into the
function. Further issues included not persisting the manifest on the
replica vBucket and not replicating snapshot markers causing
de-duplication of SystemEvents in the CheckpointManager.

Change-Id: I90db317ba560c87d226a146dddca686934989378
Reviewed-on: http://review.couchbase.org/122127
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>