MB-16239: Add support for preserving TTL [1/2]
Introduce a new frame info which allows the client to specify
that the TTL for the new item should use the TTL from the item
it replace. If no item exists the provided TTL should be used
in the newly created item.

To make reviewing the change easier the functionality is
implemented in two phases. This is phase 1 which adds
the "on the wire" spec and adds the support to subdoc
as it already does a "read, modify, write" cycle. Phase
two adds support for Set and Replace as we don't want
to change the core to always do a "read/modify/write" cycle
when the underlying engine may do this in a more efficient
way.

Change-Id: I39a960c7134687f7dec5126d3b544e3e9c8433c0
Reviewed-on: http://review.couchbase.org/106127
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>