MB-37772: Refactor ScanContext
A number of re-factoring changes around ScanContext.

1) ScanContext is now RAII in that destruction of the object will
trigger destruction of the KVFileHandle (which e.g. on couchstore,
closes the database backing the scan).

With that:

 (a) initScanContext now returns a unique_ptr to a BySeqnoScanContext
     (see 3)
 (b) scan requires a reference to a BySeqnoScanContext
 (c) destroyScanContext is now redundant and removed

2) ScanContext was constructed with shared pointers to the two callbacks
it needed, this is not required and now unique_ptr is used. The
ScanContext will take ownership of the callback objects.

 (a) A knock on effect is that the rollback interface also must specify
     a callback of the correct type.

3) ScanContext itself is given a hierarchy in preparation for by-id
scans. ScanContext is the base class (owning callbacks amongst other
data). BySeqnoScanContext is a child class which stores information
about the scan seqno range (start and end).

Change-Id: Id51a0a4bffddc10c3dd6bc0ad9d31e4bb5bfc61a
Reviewed-on: http://review.couchbase.org/122403
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>