MB-37772: Allow detection of getCollectionStats failure
Change the function to return an optional so that we can
detect a failure, such as no collection.

With this in place it lead to the detection of an issue at
warmup for collections which exist, but have not been
written to, they were getting a high-seqno of 0 when they
should get a high-seqno of their create event (startSeqno).

Change-Id: Icd8262d07cf00af6a5c5631f9836868395b2bca4
Reviewed-on: http://review.couchbase.org/122178
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>