Guard DISALLOW_COPY_AND_ASSIGN with #ifndef
An unrelated patch results in a redefinition warning because folly also
defines DISALLOW_COPY_AND_ASSIGN.

This does rely on all definitions being functionally similar, but that
seems to be a reasonable assumption given the specific macro name.

Alternatively the Folly definition could be made the canonical one, and
the definition in utility.h removed. However, the Folly header utils.h
does not solely define this macro; this solution is simpler than
potentially polluting files a large number of files with everything
from utils.h.

The Folly versions are already similarly guarded, but forcing and
relying on a specific include order seems unclean.

Change-Id: I5b63b55788cfd1b3abcba109e808dd5df9c261c7
Reviewed-on: http://review.couchbase.org/122182
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>