MB-37944: cb::const_char_buffer -> std::string_view migration prep
Second patch in preparation for replacing cb::const_char_buffer with
std::string_view.

Use the linked platform patch (which changes cb::const_char_buffer to
be a type alias to std::string_view), then fix the various remaining
issues:

* Use appropriate mutator methods to modify a std::string_view,
  instead of direct modificaiton of .buf and .len fields.

* Use explicit conversion from cb::char_buffer to
  cb::const_char_buffer (std::string_view has no such implicit
  conversion).

* Fix const-correctness issues (use cb::const_char_buffer instead of
  cb::char_buffer if the underlying data doesn't need to be changed).

* Replace some uses of `const std::string&` parameters with
  `std::string_view` to avoid creation of various std::string
  temporaries & simplify the code.

Note: Some of these changes may seem unrelated, but due to the fact
that cb::const_char_buffer is used throughout the codebase and the
differences in API c.f. std::string_view, there are quite a few
knock-on effects when changing to a new type.

Change-Id: If74f2a2943541d81b00d505ca19f6c420f22f11c
Reviewed-on: http://review.couchbase.org/122352
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>