MB-35696: Allow memcache buckets to support default collection
The approach is to first push DocKey deeper through default engine
code and then get a second Dockey (a view on the first) which drops
the collection-ID, we then hash with that key.

With that in-place the code is further updated to ensure that all
paths only accept a DocKey which is associated with the default
collection.

Further changes are made so that the engine collection management
commands now work, but only support the default collection. For
now set_collections_manifest is disabled, this may change to accept
the epoch manifest if ns_server needs that behaviour.

Change-Id: Ia8314b9146d4e1b3686d231919ab54527c4977c8
Reviewed-on: http://review.couchbase.org/122507
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>