MB-37342: Replace returns ENOENT if SW in-flight but doc does not exists
SyncReplication must not change the semantic of cmd Replace:
1) if a committed doc does not exist, return ENOENT, always
2) else, if doc exists and SW in progress, return SW_IN_PROGRESS
3) else, if doc exists and no pending SW, execute the operation

Currenlty (1) is broken. If a committed doc does not exist and a SW is
in progress, then we return SW_IN_PROGRESS. This patch fixes that.

Note: This patch also adds some engine_testapp test that requires to
introduce some missing features in the test fixture, eg:
- mad-hatter encoding for set-vbstate (required for testing SyncRepl)
- durability requirements for ::store
- support for EngineIface::store_if (required for testing cmd replace)

Change-Id: Idd0258438182996626ab68b96446de78b3db451b
Reviewed-on: http://review.couchbase.org/122702
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>