MB-37546: Update flush stats only if flush succeeds
Currently at flush-vbucket we update flush stats:
1) After persistence, for non-meta items that are flushed
2) Before persistence, for non-meta items that are deduplicated (ie, not
   flused)
3) Before persistence, for set-vbstate items

(2) and (3) are wrong as they do not consider that a flush may fail. We
fix them in this patch.

This issue has been spotted by some of the tests that stress the
flush-failure paths in EPBucket::flushVBucket. The tests are introduced
for verifying the correctness of the new backup persistrence cursor, so
the test coverage for this fix will come with the next couple of
patches.

Change-Id: Ie31c4a29c77b803b1fecf300cab5b6b100b9e0c8
Reviewed-on: http://review.couchbase.org/122392
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>