MB-37546: Do not wakeup CheckpointRemover in EPBucket::flushVBucket
This change is a pre-requirement for when we enable the backup
persistence cursor in the next patch.

When introduced, the backup persistence cursor will be removed when we
return from flushVBucekt(). Given that the CheckpointRemover must be
woken up after the backup cursor is removed, with this patch we wake
the task at Flusher, depending on the result of flushVBucket().

Apart from being a requirement for when we introduce the backup cursor,
this makes flushVBucket() more testable too.

Change-Id: Id32e56e6e589cca4ead089d3c274f088a0f6ac50
Reviewed-on: http://review.couchbase.org/122575
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>