MB-37894: Update Monotonic so assignment protects the monotonicity
The Monotonic (and WeaklyMonotonic) permitted:

Monotonic<int> a(10);
Monotonic<int> b(1);
a = b; // success in setting 10 to 1! should throw.

Because no assignment operator was defined for Monotonic that would
check the invariant.

With the assignment fixed, there are some issues in tests and code
addressed. In particular the MagmaInfo assignment in compaction
setup code tried to assign 1 = 1 on a Monotonic variable. This is
addressed by an explicit reset call which hopefully makes it more
obvious that there's more than vanilla assignment happening.

Change-Id: I0486a34a78413ad91288b34c43292d7518a081ab
Reviewed-on: http://review.couchbase.org/122654
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>