MB-37888: Observe does not transmit the key in the key
We can't use the key from the key field for RBAC as
the Observe command don't carry the key in the key field.

Given that observe don't transmit a key in the key field
it'll throw an exception when we try to create a DocKey
and terminate the connection.

Change-Id: Ib8fe58b856fcf3b5cfa528afa3a9deee6ff45f3c
Reviewed-on: http://review.couchbase.org/122686
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>