MB-38030: Keep connection open if we can for format errors
If the validator fails to execute a command due to some error
we used to send an error back to the client, and then shut
down the connection unless the command was unknown to the
system.

We don't really have to kill the connection in these cases
(unless there is a problem validating the "frame" which
could cause the system to get out of sync parsing the
input stream).

Given that clients _shouldn't_ be sending invalid encoded
commands to the server there shouldn't be any backwards
compatibility issues related to this.

Change-Id: Ib9549f0dceb2f6db9dc1b0081fe45948e0fffbe2
Reviewed-on: http://review.couchbase.org/122712
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>