MB-38031: Update vbstate json parsing to support optional fields
In the MB a specific upgrade path occurs, 4.6.5 -> 5.1.3 -> 6.5.x

Importantly, 5.1.3 is just a stepping stone to 6.5.x and no data is
written. When this occurs the vbstate (_local document) will be
missing fields which were added between 4.6.5 and 6.5.x

* hlc_epoch
* might_contain_xattrs

So these fields should not be treated as required/expected and are
optional.

This patch also makes "namespaces_supported" optional, it is an expected
field but that expectation is handled in warmup.

Change-Id: I6af7d8d432a481638c4493a78853fec76a15e460
Reviewed-on: http://review.couchbase.org/122782
Well-Formed: Wayne Siu <wayne@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>