MB-37898: Change xattr to static library
As part of upgrading Folly to C++17 we must link gflags to targets
linking Folly.

This causes problems when a mix of static and shared libraries are
used, if both link gflags - for example when trying to run
memcached_testapp:

    $ ../../memcached_testapp
    ERROR: something wrong with flag 'flagfile' in file 'gflags.cc'.
    One possibility: file 'gflags.cc' is being linked both statically
    and dynamically into this executable.

xattr is indeed one such library - it is built as a shared library and
depends on platform (which depends on folly and then gflgags), but it
is loaded into memcached_testapp which also links folly (and fglags)
statically.

Fix by changing gflags to link statically.

Change-Id: I18a00dfc97770178d32f672f87bc21ed233d681c
Reviewed-on: http://review.couchbase.org/122789
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>