Don't reduce system connections if FD limited
Currently if we have a file descriptors limit such that we need to limit
the number of connections we can have we are increasing the number of
system connections above the 5,000 default. The intent of the code is
to limit system connections to a number less than 5,000 if we have a
file descriptor limit less than 5,000, not to try to increase the
number of system connections (and decrease client connections) if we
have a limit between 5,000 and 60,000 (desired max_connections).

Change-Id: Iffbbf85723501b20df0a919da52abb3898c00356
Reviewed-on: http://review.couchbase.org/122842
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>