MB-36373: Split VBucketBGFetchItem
Due to a race condition when writes are not interlocked with compaction
we may need to BGFetch an item when compacting to ensure that we
attempt to expire only the latest version. To allow us to BGFetch this
item, compare it to the one the compactor saw, then expire it if
necessary, we need to modify the current BGFetching data structures.

Split VBucketBGFetchItem into a base class which includes the common
members for the current "FrontEnd"BGFetch and the new
"Compaction"BGFetch. Move all the current VBucketBGFetchItem uses to
the new FrontEndBGFetchItem subclass. Introduce the
CompactionBGFetchItem which we will use in a future patch to expire
items if necessary during compaction.

Change-Id: Ifd6ef2772e70265c3d4b38038b99ffbb9d1ddb93
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/122958
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>