MB-38077: Fix unity build issues
Fix a variety of issues in our source / build system which are flagged
when enabling unity build - where multiple source files are compiled
together into a single translation unit.

Most of the issues are various flavous of breaking the one definition
rule (ODR) - there can only be a single definition of every
symbol. There is two variants of this:

1. The same source files is compiled twice, which when building as
   unity build end up in the same object / link phase. The general
   solution here is to only build each source file once, adjusting
   CMakeLists.txt to ensure the targets wich need a symbol pick it up
   from the single object file.

2. Two different source files define a (static) symbol with the same
   name. This is normally fine as static symbols only have visibility
   within the same translation unit, and normally only a single source
   file is compiled at once. However this breaks down in a unitu build
   those two source files are compiled together. The general solution
   is to avoid the conflict - either by renaming one or both symbols,
   or changing their scope.

There are also some issues related to namespacing - generally a `using
namespace` directive in a .cc file is a safe thing to do, however with
a unity build that can "pollute" other source files included in the
translation unit after the first. This can be avoided thier by
removing the `using namespace` or adding explicit namespacing
prefixing (e.g. ::to_string to refer to a symbol in the global
namespace).

See also
https://onqtam.com/programming/2019-12-20-pch-unity-cmake-3-16/ which
has a good writup on getting unity builds working.

Change-Id: Ie530d7e061e501176a95a359146b05165b8ef3a1
Reviewed-on: http://review.couchbase.org/122939
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>