MB-38031: Don't bring up the bucket when couchstore file cannot be read
In MB-38031 the json parsing error lead to the bucket coming online
with no data loaded (and incorrect high-seqno in memory). Left in
this state we would of accepted writes which could lead to further
problems.

This commit adds an exception into the CouchKVStore constructor which
is thrown when we have a vb file on disk which we cannot open or cannot
read VB state (a json issue for example). This exception is caught and
will cause create bucket to fail meaning we cannot bring the broken
bucket online.

Note: the CouchKVStore construction occurs inside engine initialize,
which is wrapped in a try...catch inside memcached.cc and now in
engine_testapp, both cases when an exception is caught will call
engine destroy to clean-up.

Change-Id: I4e4edd731e7546133caf37eab6f6cad5cb8b4141
Reviewed-on: http://review.couchbase.org/122854
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>