MB-36373: Refactor evp_store unit tests
The evp_store unit tests currently run only for couchstore, and some
combination of eviction policies. Refactor this test suite to use (where
possible) the STParameterizedBucketTest fixture so that we can run the
tests for non-couchstore backends.

Some tests have their own parameterization including bloom filters. Add
a new testsuite for these as GTest does not allow you to have multiple
hierarchies. Alternatively we could add a bloom filter on/off to the
fullEviction flag and re-use the STParameterizedBucketTest.

Change-Id: I12035d7c920fd84b34e2ca9ad1b82f0018783a54
Reviewed-on: http://review.couchbase.org/123085
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>