MB-36373: BGFetch non-resident expiries for magma
Magma does not interlock writes and compaction so a race condition
exists where a compaction can find and attempt to expire an item that
has been logically superseded on disk. If the new item is still resident
in the HashTable then the current code works fine, however, if the item
is not resident then we cannot blindly expire the item that the
compactor has passed to us. Instead, we need to BG fetch the item from
disk and only expire it if it's still the latest version.

Change-Id: Ia1aed5a83467ffa8dd4f7a83e53439011ea9a1e5
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/122960
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>