MB-38019: Improve manifest update error checking
A number of related changes that begin with changing the manifest
update function to return a status code which describes any issue.

Two errors are currently defined.

1) Behind - an update was attempted and the input manifest as a UID
which is "behind" the current UID.

2) EqualWithDifferences - an update was attempted, the input as the
same UID but result in an attempt to drop/create a scope/collection,
i.e. the manifest is not equal.

We now use these errors to better handle some situations, particularly
the issue in the MB where we wish to ignore the "Behind" error that
can occur in vbucket promotion (replica to active).

Change-Id: I2bce0b3d082b3e6a86b715c6297a8a02ab52d873
Reviewed-on: http://review.couchbase.org/123288
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>