MB-38197: On-disk #items incorrect after Delete + SyncAdd
If an item is deleted, but re-created via a SyncAdd _before_ the
Delete is persisted to disk, then the onDiskTotalItems is incorrectly
not decremented. This results in over-counting of how many items exist
in the VBucket for Full Eviction buckets.

The bug is caused by an incorrect assumption in
VBucket::deletedOnDiskCbk() - that onDiskTotalItems should only be
decremented if a StoredValue was found in the HashTable.

This was a correct assumption prior to SyncWrites - a dirty, delete
StoredValue would never be completely removed from the HashTable - at
most it could be "UnDeleted" and replaced with a non-deleted
StoredValue.

However, SyncWrites can replace a dirty, deleted SV with a Pending
StoredValue, and given Pending SVs are not included in
onDiskTotalItems then the decrement was skipped.

Change-Id: I98bfe37514f22ebd0749f6044c04979748ee80f2
Co-authored-by: Dave Rigby <daver@couchbase.com>
Reviewed-on: http://review.couchbase.org/123472
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>