Minor refactor of doScopeStats and doCollectionStats
Small reordering to favour early exits (in accordance with the coding
standards) to be consistent with and to prepare for a following patch.

Among other minor changes, move to using std::string_view to avoid
string copies, and try to limit nesting by preferring early exits in
accordance with the coding standards.

Change-Id: Ic1841ecacc52b0159bcfc2d17274a50e91fb2184
Reviewed-on: http://review.couchbase.org/123317
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>