MB-37948: Use the proposedVBState when checking state
proposedVBState was added to the master branch as part of fixing
MB-36771, when a set-vb-state item is flushed it is this variable which
gets the new vbucket state.

* http://review.couchbase.org/#/c/117709/

A merge from mad-hatter changes part of the flusher which was reading
the vb state from the VBucket object to read from a flusher 'vbstate'
object.

* http://review.couchbase.org/#/c/122273/-1..1

When the two are combined, the code reading the 'vbstate' now finds it
never gets updated with the new vbucket state and hence we see the bug.

The bug is that the vbucket is active, yet a state change occurs (to
replica) followed by the write of 5 items. Depending on the timing of
the flusher we may flush the 5 items and the set-vb-state, or flush them
separately.

When flushed as a single batch, 'vbstate' begins as active and as the
flusher progresses through the batch it finds the set-vb-state item and
then writes the new state of replica to proposedVBState. Once the batch
is finished the clause for calculating the snapshot range for an active
vbucket is incorrectly entered and the range is set to {5,5} -> fail.

Change-Id: Iae98de4e69749653a26d923bd3b99d039840ee3d
Reviewed-on: http://review.couchbase.org/123329
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>