MB-38173: Track key allocations as part of checkpoint overhead
Currently we are tracking checkpoint memory overhead as size of the
Checkpoint plus the amount of memory allocated in the keyIndex map plus
the amount of memory in the toWrite queue. If we have expelled all
items then the only substantial memory allocation will be the keyIndex.

An oversight was made when adding the tracking of the keyIndex such
that we do not track any heap memory allocations of keys
(std::strings). Add an "AllocatorTracked"DocKey to allow us to track
the heap allocations of keys in the checkpoint indexes.

Change-Id: I184d6be8ef1bfb527011201acbfbee31e46c5f2a
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/123313
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>