MB-37798: Read vbstate from disk for scans
MB-35631 changed when cachedVBState gets updated so the latest vbstate
is found on disk, not in the cache. Scans should use the latest
vbstate. To fix this, we need to add a new function that reads vbstate
from disk without loading the cache. To do this, we'll rename the
existing function, readVBStateFromDisk to loadVBStateCache and repurpose
the old function to just reading the vbstate from disk.

The test added reflects the current broken state of dcp scans which
don't retain the scan snapshot. That will be fixed in a subsequent
checkin.

Its unclear how MB-35631 did not cause a tsan race condition but this
patch fixes that issue.

Change-Id: Ieca302cb8adbbc2e44f80f9ccfa360f748f4c27e
Reviewed-on: http://review.couchbase.org/123509
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>