MB-37798: Provide consistent scan snapshot
When a scan encounters an ENOMEM, we don't restart the scan, we try to
continue from where we left off. This change moves the magma seq index
iterator into the ScanContext in order to remain on the same magma
snapshot to provide a consistent snapshot view for the entire scan.

Change-Id: I1ccdad2469728dba1ae55c318b025b25244f6452
Reviewed-on: http://review.couchbase.org/123513
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>